requirements.txt
torch>=1.1
numpy
hydra-core>=1.0.0
pytorch-lightning
tqdm